requirements.txt
# MCP Figma Sync Server Requirements
# Yêu cầu cho MCP Figma Sync Server

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP Client
aiohttp==3.9.1
requests==2.31.0

# Configuration
pydantic==2.5.0
pydantic-settings==2.1.0

# Git Integration
GitPython==3.1.40

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Logging
structlog==23.2.0

# Optional: Database (SQLite for now)
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Optional: Caching
redis==5.0.1
aioredis==2.0.1